Added: 4-way player movement
Hey, i added 4-way player movement.

for more of the smaller stuff i tweaked:

added a global light and a spotlight for basic lighting to the scene. as well as added some sorting layers and the demo player( square sprite).

if you are worrying i added a triangle in the background for light showcasing and i am going to research the bug for the player colliding with the light. =)